requirements.txt
certifi==2016.2.28
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
honcho==1.0.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pysqlcipher3==1.0.2
python-dateutil==2.6.1
six==1.11.0
SQLAlchemy==1.1.14
uWSGI==2.0.15
Werkzeug==0.12.2
WTForms==2.1
